+++ src/connectedk8s/azext_connectedk8s/_clientproxyutils.py
@@ -24,8 +24,6 @@ from psutil import (
     net_connections,
 )
 from knack.log import get_logger
-from azure.cli.core.auth.msal_authentication import ServicePrincipalCredential
-
 logger = get_logger(__name__)
 
 
@@ -115,17 +113,9 @@ def fetch_and_post_at_to_csp(cmd, api_server_port, tenant_id, kid, clientproxy_p
     token_data = {"token_type": "pop", "key_id": kid, "req_cnf": req_cnf}
     profile = Profile(cli_ctx=cmd.cli_ctx)
     try:
-        credential, _, _ = profile.get_login_credentials(
-            subscription_id=profile.get_subscription()["id"],
-            resource=consts.KAP_1P_Server_App_Scope,
-        )
-        if isinstance(credential._credential, ServicePrincipalCredential):
-            # This is a workaround to fix the issue where the token is not being refreshed
-            # https://github.com/AzureAD/microsoft-authentication-library-for-python/pull/692
-            credential._credential.remove_tokens_for_client()
-        accessToken = credential.get_token(
-            consts.KAP_1P_Server_App_Scope, data=token_data
-        )
+        credential, _, _ = profile.get_login_credentials(subscription_id=profile.get_subscription()["id"],
+                                                         resource=consts.KAP_1P_Server_App_Scope)
+        accessToken = credential.get_token(consts.KAP_1P_Server_App_Scope, data=token_data)
         jwtToken = accessToken.token
     except Exception as e:
         telemetry.set_exception(