+++ src/image-copy/azext_imagecopy/tests/latest/test_image_copy_commands.py
@@ -6,11 +6,13 @@
 import time
 from azure.cli.testsdk import ScenarioTest, ResourceGroupPreparer, live_only
 from azure.cli.testsdk.constants import AUX_SUBSCRIPTION
+from azure.cli.testsdk.scenario_tests import AllowLargeResponse
 
 
 class ImageCopyTests(ScenarioTest):
 
     @live_only()
+    @AllowLargeResponse(size_kb=9999)
     @ResourceGroupPreparer(name_prefix='cli_test_image_copy_', location='westus')
     def test_image_copy(self, resource_group):
         self.kwargs.update({
@@ -20,27 +22,31 @@ class ImageCopyTests(ScenarioTest):
             'image': 'image1',
             'hyperVGeneration': 'V2',
             'subscription2': AUX_SUBSCRIPTION,
+            'vnet': 'vnet',
+            'subnet': 'subnet'
         })
 
-        self.cmd('vm create -g {rg} -n {vm} --image canonical:0001-com-ubuntu-server-focal:20_04-lts-gen2:latest'
-                 ' --admin-username clitest1 --generate-ssh-key')
+        self.cmd(
+            'vm create -g {rg} -n {vm} --admin-username theuser --image OpenLogic:CentOS:7.5:latest --admin-password testPassword0 '
+            '--authentication-type password --nsg-rule NONE --subnet {subnet} --vnet-name {vnet}')
+        self.cmd('network vnet subnet update -g {rg} --vnet-name {vnet} -n {subnet} --default-outbound-access false')
         self.cmd('vm run-command invoke -g {rg} -n {vm} --command-id RunShellScript --scripts '
                  '"echo \'sudo waagent -deprovision+user --force\' | at -M now + 1 minutes"')
         time.sleep(70)
         self.cmd('vm deallocate -g {rg} -n {vm}')
         self.cmd('vm generalize -g {rg} -n {vm}')
-        self.cmd('image create -g {rg} -n {image} --source {vm} --hyper-v-generation {hyperVGeneration}')
+        self.cmd('image create -g {rg} -n {image} --source {vm}')
         self.cmd('group create -g {rg2} -l eastus')
         self.cmd('image copy --source-object-name {image} --source-resource-group {rg} --target-location eastus '
                  '--target-resource-group {rg2} --target-name {image} --target-subscription {subscription2} '
                  '--temporary-resource-group-name {temporary_rg} --cleanup --only-show-errors')
         self.cmd('image show -g {rg2} -n {image} --subscription {subscription2}', checks=[
             self.check('name', '{image}'),
-            self.check('hyperVGeneration', '{hyperVGeneration}'),
         ])
 
     # the recoding file does not contain the image copy command, so that this test have to be run in live mode
     @live_only()
+    @AllowLargeResponse(size_kb=9999)
     @ResourceGroupPreparer(name_prefix='cli_test_image_copy_loc_', location='westus')
     def test_image_copy_locations(self, resource_group):
         self.kwargs.update({
@@ -49,10 +55,13 @@ class ImageCopyTests(ScenarioTest):
             'rg2': self.create_random_name(prefix='cli_test_image_copy_loc2_', length=30),
             'loc1': 'southeastasia',
             'loc2': 'australiaeast',
+            'vnet': 'vnet',
+            'subnet': 'subnet'
         })
 
         self.cmd('vm create -g {rg} -n {vm} --admin-username theuser --image OpenLogic:CentOS:7.5:latest --admin-password testPassword0 '
-                 '--authentication-type password --nsg-rule NONE')
+                 '--authentication-type password --nsg-rule NONE --subnet {subnet} --vnet-name {vnet}')
+        self.cmd('network vnet subnet update -g {rg} --vnet-name {vnet} -n {subnet} --default-outbound-access false')
         time.sleep(70)
         self.cmd('vm deallocate -g {rg} -n {vm}')
         self.cmd('vm generalize -g {rg} -n {vm}')