+++ src/communication/HISTORY.rst
@@ -2,6 +2,16 @@
 
 Release History
 ===============
+1.11.0
+++++++
+* Update to email SDK version to 1.0.1b1.
+* Adding new feature inline attachments to Email communication send mail.
+
+1.10.0
+++++++
+* Adding new parameter waituntil to Email communication send mail.
+* Adding new command get send status to Email communication.
+
 1.9.3
 ++++++
 * Update Email service create - Global as the default value for Location